ebi-ait/dcp-ingest-central#142: used flask blueprint for geo accession route
